Bugfix - only try to delete Flux extension when it is present (in failed state)

+++ azure_jumpstart_ag/artifacts/PowerShell/AgLogonScript.ps1
@@ -1112,7 +1112,11 @@ foreach ($resource in $resources) {
 
                 try {
 
-                    az k8s-extension delete --name "flux" --cluster-name $resourceName --resource-group $Env:resourceGroup --cluster-type $ClusterType --force --yes
+                    if ($extension) {
+
+                        az k8s-extension delete --name "flux" --cluster-name $resourceName --resource-group $Env:resourceGroup --cluster-type $ClusterType --force --yes
+
+                     }
 
                     az k8s-extension create --name "flux" --extension-type "microsoft.flux" --cluster-name $resourceName --resource-group $Env:resourceGroup --cluster-type $ClusterType --output json | ConvertFrom-Json -OutVariable extension
